Add neighborhood and subregion fields to PlaceQuery.
In order to better support geocoding outside the US, add the neighborhood
(sub-municipal) and subregion (sub-state) fields to the PlaceQuery()
class. These fields are primarily used by the EsriWGS geocoder. Note that
the Esri geocoders which use endpoints at tasks.arcgis.com  have not been
updated because they are deprecated and Esri plans to deactivate them at
the end of 2013.