Add Esri magic key support.
Esri returns search suggestions for freeform text using structure and endpoint
documented at
https://developers.arcgis.com/rest/geocode/api-reference/geocoding-suggest.htm

The results of this are a string and a key that appears to be a uuid for the
location.

The find endpoint documented at
https://developers.arcgis.com/rest/geocode/api-reference/geocoding-find.htm
accepts the "magicKey" returned from the results of the suggest endpoint to
quickly find a specific location. In testing it was determined that the
actual search text is ignored when there is a magicKey though it is still
required in order to get any results.

The allows the PlaceQuery to supply a "key" from kwargs that will be used in the
find query in EsriWGS and the SSL version of it.