Support uuid/magic_keys.
Mapzen has a different endpoint for known places, this is similar to the
ESRI magic key in that it can be used to find an exact location by id
and works in conjunction with the Mapzen autocomplete endpoint. We can use
the same 'key' identifier in the PlaceQuery to pass to the geocoder and
just switch the endpoint if this is set.