Support custom mapzen urls.
Mapzen, being a fully open source geocoder, can be spun up on any server
at any URL. This allows for users to create their own mapzen/pelias
server and use it as a live geocoder. We support this feature here by
making the api key optional and allowing for the instance URL to be
overridden by providing an url in the settings dictionary passed to the
constructor.