Move flake8 installation from setup.py to .travis.yml
Flake8 shouldn't be required for production installation, but putting it
in `setup_requires` causes it to be installed (or to crash while trying
to install, in some cases).
Just installing it and running it within the Travis config avoids any
problems around around how to specify it in setup.py.