Reconcile string types between regex and response keys
The regex for preprocessor filtering can be either unicode or bytestring,
and the values being filtered have to match. So this gets the type and
casts the strings being tested to it.

This also changes the initial type check for the regex argument to be more
complete and correct across Python versions.