Restore type checks for different string types
In a few places, `2to3` converted `type(foo) in (str, unicode)` to
`type(foo) in (str, str)`, which is not very useful. It also will
fail if you run it in Python 2 and the type is `unicode`.
This makes the type check work in Python 2 and Python 3, using a
workaround that doesn't require depending on `six` or `future`.