Fix template string for AttrList{In|Ex}cludes.__repr__
These classes were based on `AttrFilter` and `AttrExclude` but don't
have the `exact_match` property. So the template string in their __repr__
method was causing a crash. It only comes up if you try to get the
string representation of a geocoder object, so pretty much only while
debugging, but it's a bit confusing and unhelpful when it happens.