Use explicit encoding in .decode()
As noted by @ddohler on PR #61 (https://github.com/azavea/python-omgeo/pull/61#discussion_r334910782),
the default encoding in Python 2 is `ascii`, so `.decode()` is going to
crash on anything with non-ASCII characters in it. It might not actually
be possible to reach this line in Python 2, and in Python 3 the the default
is "utf-8", but no harm in handling it better if it somehow happens.