Tolerate missing Loc_name, warn if on missing keys
Previously we assumed that Loc_name would always be provided, given
that we ask for it in the requested field list. However, when using
the magicKey, the Loc_name is not provided, causing geocoding to
fail. By making it optional, we allow those geocoding requests to
succeed.

Also, add a WARNING to the log when there's a missing key, instead
of just silently passing, for better observability.